Added functionality for drawing tree lines (node connectors).
Overview of the algorithm:
As the current drawing algorithm uses DFS, this commit adds a preliminary BFS that precedes the DFS;
Its purpose is to supply a mapping from each node to a boolean that states whether there exists a sibling(for the current node) that will be printed next (can be done by remembering the previous scanned node and comparing the parent pointer and depth level, at each node visit).
This information will be later be used by the DFS, to maintain (while DFSing) a mapping from depth D (D < depth of current node) to a boolean that states whether there's another sibling node to print in that depth. If for depth i (i < depth of current DFS node), then a vertical line needs to be printed.

For instance:
given the following tree:

root
+-- child 1
|   +-- child1.1
|   +-- child1.2
+-- child 2
    +-- child2.1

The result of step 1 (the preliminary BFS) will be:
{child1: true, child1.1: true, child1.2: false, child2: false, child2.1: false}

Then, step2 (printing with DFS) can use this mapping to maintain the following array (next to each node is the array state, when visiting this node in the DFS):

root []
+-- child 1 [False (root does not have any siblings)]
|   +-- child 1.1 [False, True (child2 exists as the next printed sibling of child1)]
|   +-- child 1.2 [False, True (child2 exists as the next printed sibling of child1)]
+-- child 2 [False]
    +-- child 2.1 [False, False (no sibling of child2 is going to be printed next)]
The above array indicates, for each index, whether to print a horizontal line at that indent depth, for the current node.

Today i learned: 'using' keyword (same as typedef)